Add Version class
Follows SemVer specification, parsing a version string to extract
major, minor, patch, prerelease type and version, and build
metadata. It also includes operator overloads for the assignment
and comparison operators, allowing comparison not only with other
Version objects, but also with strings.

For information regarding the format in which version strings are
parsed, see the SemVer specification: https://semver.org/