Add range class
Allows a set of comparators to be specified and tested to see if
they are satisfied by a particular version. Comparator sets are
divided by ||s - logical ORs. Each comparator set is made up of
a number of comparators, divided by spaces. The comparator is
composed of an operator and a version, with the operator coming
first. Only the following operators are implemented: <, <=, >, >=,
= (if no operator is specified, = is assumed).